+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-## [Unreleased]
+## [0.12.5] - 2024-04-15
+
+- Restrict validation rules for Hash type objects
+
+## [0.12.4] - 2024-04-14
+
+- Revert 0.12.2 changes
+
+## [0.12.3] - 2024-04-14
+
+- Default option for Integer type should support empty string as well
 
 ## [0.13.0] - 2024-04-11